% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subfunction.R
\name{Estimate.EDR.from.pilot}
\alias{Estimate.EDR.from.pilot}
\title{Predict EDR}
\usage{
Estimate.EDR.from.pilot(Data, status, group.name = c("Control", "Case"), FDR,
  M, target.N, target.R = NULL, target.theta = NULL, tol = 0.1,
  filter = T, filter.level = 5)
}
\arguments{
\item{Data}{Input pilot data.}

\item{status}{A vector of group labels.}

\item{group.name}{A vector of length two. First element is the group name of first group, and the second element is of second group.}

\item{FDR}{FDR level.}

\item{M}{Number of iterations for parametric bootstrap.}

\item{target.N}{Targeted sample size.}

\item{target.R}{Targeted sequencing depth.}

\item{target.theta}{Targeted proportion of sample size in second group compared to first group.}

\item{tol}{Threshold for deciding if use CDD to estimate the proportion of DE genes.}

\item{filter}{Filter genes based on mean counts?}

\item{filter.level}{Filter the genes with mean counts less than this level.}
}
\value{
An object of class list is returned:
Result is a list of length M (depends on how many times of parametric bootstrap you run), and each component is also a list of length of variable target.R.
Each component contains a matrix with target N in row and summary statistics in columns. Summary statistics provided are TP, TN, EDR.
}
\description{
Predict EDR based on a pilot data.
}
\author{
Chien-Wei Lin
}
