% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subfunction.R
\name{Estimate.true.EDR}
\alias{Estimate.true.EDR}
\title{Calculating true EDR}
\usage{
Estimate.true.EDR(Data, status, group.name = c("Control", "Case"), is_DE, FDR,
  filter = T, filter.level = 5)
}
\arguments{
\item{Data}{Input pilot data.}

\item{status}{A vector of group labels.}

\item{group.name}{A vector of length two. First element is the group name of first group, and the second element is of second group.}

\item{is_DE}{A vector of True or False which indicate the DE status of each gene.}

\item{FDR}{FDR level.}

\item{filter}{Filter genes based on mean counts?}

\item{filter.level}{Filter the genes with mean counts less than this level.}
}
\value{
An object of class list is returned:
The object contains a vector of three summary statistics, which are TP_true, TN_true, and EDR_true
Each component contains a matrix with target N in row and summary statistics in columns. Summary statistics provided are TP, TN, EDR.
}
\description{
Calculating true EDR based on a given data with given true DE information.
}
\author{
Chien-Wei Lin
}
