% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subfunction.R
\name{Predict.EDR}
\alias{Predict.EDR}
\title{Predict EDR from surface fitting}
\usage{
Predict.EDR(target.N, target.R, Surface.para)
}
\arguments{
\item{target.N}{Target sample size.}

\item{target.R}{Target sequencing total reads.}

\item{Surface.para}{Parameters of two dimensional EDR surface, estimated from function Summarize.RNASeqDesign.}
}
\value{
Predicted EDR.
}
\description{
Predict EDR at any sample size and sequencing depth after two-dimensional surface fitting.
}
\author{
Chien-Wei Lin
}
