% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subfunction.R
\name{Summarize.RNASeqDesign}
\alias{Summarize.RNASeqDesign}
\title{Summary EDR result and surface fitting}
\usage{
Summarize.RNASeqDesign(Result, target.N, target.R)
}
\arguments{
\item{Result}{Object from the output of function Estimate.EDR.from.pilot.}

\item{target.N}{Target sample size used to predict EDR.}

\item{target.R}{Target total sequencing reads used to predict EDR.}
}
\value{
An object of class list is returned:
The object contains a matrix of predicted EDR, where the rows are different target N and columns are different target R, and a vector of four parameters from two-dimensional surface fitting.
}
\description{
Summarize the result from function Estimate.EDR.from.pilot and output the summarized EDR table and estimated parameters of two-dimensional surface fitting.
}
\author{
Chien-Wei Lin
}
